+++ src/main/java/com/tiki/server/memberteammanager/service/MemberTeamManagerService.java
@@ -15,7 +15,4 @@ public class MemberTeamManagerService {
 
     MemberTeamManagerFinder memberTeamManagerFinder;
 
-    public BelongTeamsResponse showBelongTeam(long memberId){
-        val belongTeams = memberTeamManagerFinder.findBelongTeamById;
-    }
 }
