+++ src/main/java/com/tiki/server/team/adapter/TeamFinder.java
@@ -3,11 +3,15 @@ package com.tiki.server.team.adapter;
 import static com.tiki.server.team.message.ErrorCode.INVALID_TEAM;
 
 import com.tiki.server.common.support.RepositoryAdapter;
+import com.tiki.server.memberteammanager.dto.response.BelongTeamsResponse;
 import com.tiki.server.team.entity.Team;
 import com.tiki.server.team.exception.TeamException;
 import com.tiki.server.team.repository.TeamRepository;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.jpa.repository.Query;
+
+import java.util.Optional;
 
 @RepositoryAdapter
 @RequiredArgsConstructor
@@ -18,5 +22,9 @@ public class TeamFinder {
 	public Team findById(long teamId) {
 		return teamRepository.findById(teamId)
 			.orElseThrow(() -> new TeamException(INVALID_TEAM));
+	}
+
+	public Optional<Team> findBelongTeamByMemberId(long memberId){
+		return teamRepository.findBelongTeamByMemberId(memberId);
 	}
 }
